Illustration - Problem with txn_doublespend.py
This small change to the chosen test amounts makes the test
produce unpredictable output.  It produces one of these outputs
for each run, with apparent 50% probability:

Output 1:
1260.00000000
-1220.00000000
100.00000000

Output 2:
10.00000000
30.00000000
100.00000000

The problem is that the test currently relies on SyncMetaData,
which has unpredictable results when invoked for non-clone
conflicts.

In this illustration, the double spend invalidates both test
transactions, so SyncMetaData is called twice.  The from-
account of whichever synced-to transaction is processed last
wins.

This commit will need to be squashed before merge.